We can check our code for errors with webpack too!
> npm install eslint-loader eslint --save-dev

We'll need to grab my `.eslintrc` file. I recommend looking through this file, but for now just drop it in the root of your project (provide URL on slide)

Add eslint as a pre-loader - this will run eslint before any other build steps - if the check fails, the rest of the build won't run.

Run webpack and fix the eslint errors.
  - Re-running `webpack` is annoying - they
    give us a nice option - `webpack --watch`