Running a local server
> npm install gulp-server --save-dev

Create "webserver" task, with "./" folder. Illustrate livereload/watching for CSS and for javascript.

Illustrate an ESLint error. Oh no, the server stopped! We can remove `failAfterError()` and this doesn't happen any more, but if we had a task to build for production we would want to fail on ESLint error. We could make 2 different tasks, "lint" and "lint-strict" maybe? Several ways to solve the problem.